Make it work with updated Simply Static plugin
Since December 2020 there is life again in the original Simply
Static plugin. This means we do not need to use our forked version
for php > 7.2. Also another bug we fixed in our fork is fixed in
the updated original plugin.

But because the custom class loader of the original plugin now only
is required after all plugins are loaded. We need to use another hook
that will be fired after the classes are loaded.